wip checking remote and local mode
